gotiny: Move to own pkg
